requirements.txt
discord.py
aiohttp
Pillow
pyLast
yt-dlp